sensor code added; integration wip

+++ src/sample.h
@@ -8,7 +8,7 @@ class Sample
   public:
   Sample(T init_val) : head(0)
   {
-    for (size_t i=0; i < N; i++) {
+    for (unsigned int i=0; i < N; i++) {
       samples[i] = init_val;
     }
   }
@@ -22,7 +22,7 @@ class Sample
     }
   }
 
-  size_t items_within_range(T low, T high, int history = N)
+  unsigned int items_within_range(T low, T high, int history = N)
   {
     if (history > N) {
       history = N;
@@ -33,8 +33,8 @@ class Sample
       idx = N - 1;
     }
 
-    size_t iwr = 0;
-    for (size_t i = 0; i < (size_t) history; i++) {            
+    unsigned int iwr = 0;
+    for (unsigned int i = 0; i < (unsigned int) history; i++) {            
       if (low <= samples[idx] && samples[idx] <= high) {
         iwr += 1;
       }
@@ -55,7 +55,7 @@ class Sample
 
     // copy samples to a temporary array for median algorithm (modifies in place)
     T temp_samples[N];
-    for (size_t i = 0; i < history; i++) {
+    for (unsigned int i = 0; i < history; i++) {
       temp_samples[i] = samples[i];
     }
 
@@ -63,14 +63,14 @@ class Sample
   }
 
   // for debugging/testing
-  T get_item(size_t idx) const { return samples[idx]; }
+  T get_item(unsigned int idx) const { return samples[idx]; }
 
-  size_t size() const { return (size_t) N; }
+  unsigned int size() const { return (unsigned int) N; }
 
   private:
   T samples[N];
   
   // head is the offset of the next place to populate in the ring buffer
-  size_t head;
+  unsigned int head;
 
 };
